Trait livesplit_wrapper::HostFunctions

pub trait HostFunctions {
    fn print(&self, str: &str) { ... }
    fn attach(&self, name: &str) -> Option<Process> { ... }
    fn start(&self) { ... }
    fn pause(&self) { ... }
    fn unpause(&self) { ... }
    fn split(&self) { ... }
    fn reset(&self) { ... }
    fn set_game_time(&self, time: f64) { ... }
    fn set_tick_rate(&self, rate: f64) { ... }
    fn state(&self) -> TimerState { ... }
    fn set_variable(&self, key: &str, value: &str) { ... }
}

The autosplitter’s interface for interacting with the LiveSpilit timer.

Provided Methods

fn print(&self, str: &str)

Output a message. This can be used for debugging and/or sending error messages to the player through whichever LiveSplit frontend they’re using. Note that because autosplitters run in WASM, they don’t have access to STDOUT or files, so typical solutions like println! and logging will not work (this could chagne in the future as LiveSplit plans to support WASI).

fn attach(&self, name: &str) -> Option<Process>

Attach to a process running on the same machine as the autosplitter.

fn start(&self)

Start the timer for a run. Note that this will silently do nothing on subsequent calls. To start a new run, call reset() and then start().

fn pause(&self)

Pause the game time counter. This is often used when entering a loading screen or end level screen for games that use in game time rather than real time. It may be a good idea to call set_game_time() immediately after pausing so that LiveSplit’s game time counter shows the exact current time.

fn unpause(&self)

Resume the game time counter. Note that

fn split(&self)

Mark the current split as finished and move to the next one.

fn reset(&self)

Reset the run. Don’t do this automatically when a run has finished, and in general be conservative about resetting runs from the autosplitter. Common practice is to do so only if there’s an unambiguous signal that the player is done with this run.

fn set_game_time(&self, time: f64)

Set the game time. Note that if the timer is not paused, the time shown will keep incrementing immediately after it is set to the given value.

fn set_tick_rate(&self, rate: f64)

Set the rate at which the update function will be called (in Hz).

fn state(&self) -> TimerState

Get the current state of the timer. This is how the autosplitter can detect if the player manually paused or reset a run.

fn set_variable(&self, key: &str, value: &str)

Set a variable which can be displayed by LiveSplit. This is commonly used for features like death counters.